feat(extension): FILL/TYPE 写入机制转正为 CDP Input.insertText
对标 playwright-mcp/chrome-devtools-mcp/Stagehand:fill/type 默认走 CDP
Input.insertText(isTrusted=true)。spike(reports/spike-cdp/)fill 矩阵实测
cdpFill 无回归 + 3 保真优势(maxlength 被尊重/beforeinput 触发/isTrusted),
value-setter/page-side-dispatch 降为 forceSynthetic + CDP attach 失败的降级。

实现(单一探测真源,写入机制分叉,非各处补 if):
- FILL:厚探测(族 B 类型分流 select/checkbox/contenteditable→INVALID_TARGET)
  提取为单一 inline func fillProbeOrWrite,prepareOnly=true 探测+focus+全选
  不写值(CDP insertText 接手)、prepareOnly=false 走 value-setter(族 F 受控
  绕过+族 A NO_EFFECT 保留作降级)。fill-reject 分流照旧前置,对两路都生效。
- TYPE:select-all 从 cdpType 实验分支提到统一 probe(input/textarea el.select()
  / contentEditable range);input/textarea 默认走 CDP,attach 失败降级 dispatch。
- 共同:元素级错误直抛、CDP readback 空→NO_EFFECT 直抛(降级同果)、
  forceSynthetic 纯降级逃生口、cdpFill/cdpType 实验参数退役(默认即 CDP)。
- inline func 经 executeScript 注入不引用模块级 helper(本 spike P0 教训)。

删 spike-cdp-fill-type 脚手架,新 fill-type-cdp-first.test.ts 13 case 锁转正
契约;3 处 source-grep 适配新结构(契约未变)。ext 958/958、build 通过。
